updated auth_failed method

+++ Project03/routers/todos.py
@@ -17,6 +17,11 @@ def get_db():
     finally:
         db.close()
 
+def auth_failed(user):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail="Could not validate credentials")
+
 db_dependancy = Annotated[Session, Depends(get_db)]
 user_dependancy = Annotated[dict, Depends(get_current_user)]
 
@@ -30,16 +35,12 @@ class TodoRequest(BaseModel):
 
 @router.get("/", status_code=status.HTTP_200_OK)
 async def read_all(user: user_dependancy, db: db_dependancy):
-    if user is None:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
-                            detail="Could not validate credentials")
+    auth_failed(user)
     return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
 async def read_todo(user: user_dependancy, db: db_dependancy, todo_id: int = Path(gt=0)):
-    if user is None:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
-                            detail="Could not validate credentials")
+    auth_failed(user)
     todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
@@ -49,10 +50,7 @@ async def read_todo(user: user_dependancy, db: db_dependancy, todo_id: int = Pat
 async def create_todo(user: user_dependancy,
                       db: db_dependancy,
                       todo_request: TodoRequest):
-    if user is None:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
-                            detail="Could not validate credentials")
-
+    auth_failed(user)
     todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id'))
 
     db.add(todo_model)
@@ -63,10 +61,7 @@ async def update_todo(user: user_dependancy,
                       db: db_dependancy,
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
-
-    if user is None:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
-                            detail="Could not validate credentials")
+    auth_failed(user)
     todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
@@ -80,12 +75,12 @@ async def update_todo(user: user_dependancy,
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependancy, todo_id: int = Path(gt=0)):
+async def delete_todo(user: user_dependancy, db: db_dependancy, todo_id: int = Path(gt=0)):
     todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-
+    auth_failed(user)
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found.")
     
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
 
     db.commit()
